Adding the undocumented Host and Name tags
These are referenced here:

https://groups.google.com/forum/#!msg/tumblr-themes/v3SxNAW70cU/dHCZWn7yFPYJ

Specifically:

They are Undocumented.

{Name}  Gives the "Tumblr Name" of the user
{Host}  Gives the "Host name" for the blog  (ie, if the URL is http://foobar.tumblr.com,  this generates "foobar.tumblr.com")

Note that for Tumblrs with a Custom domain name, the {Host} variable will return that custom domain name. So if the Tumblr's URL is "http://www.my-awesome-tumblr.com",  {Host}  generates "www.my-awesome-tumblr.com"

The {Name} variable always returns the actual Tumblr Name of the blog, regardless of whether it has a custom domain.